Refactor view-to-svg, implement baseline wireframe mode for views
