+++ src/svg/util.ts
@@ -1,0 +1,12 @@
+const attributes = (settings) => {
+  let attributeString = ""
+  for (const key in settings) {
+    if (settings.hasOwnProperty(key)) {
+      const element = settings[key]
+      attributeString += ` ${key}="${element}"`
+    }
+  }
+  return attributeString
+}
+
+export const $ = (type, settings) => `<${type}${attributes(settings)}/>`
